requirements_test.txt

# Pre-commit requirements
bandit==1.6.2
black==20.8b1
codespell==2.0.0
flake8-docstrings==1.5.0
flake8==3.8.4
isort==5.6.4
pydocstyle==5.1.1
pyupgrade==2.7.4
yamllint==1.24.2


# Unit tests requirements
pre-commit==2.9.3
pylint==2.6.0
pytest<6.0
pytest-cov<3.0.0
pytest-homeassistant-custom-component